+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Date: 2022-11
 * Added new "never-ran" e-mail template (issue #63).
 * Added the ability for users to configure the e-mail regular expression in slurm-mail.conf (issue #54).
 * Added support for optionally running commands after submitting jobs when running integration tests.
+* Added `Date` and `Message-ID` e-mail headers (issue #64).
 * Updated GitHub workflows to remove use of `set-output` following [deprecation](https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands).
 
 Version 4.0